ai_service/requirements.txt
flask
ultralytics
pillow
easyocr
flask-cors
